requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Transformers and ML
simpletransformers>=0.70.0
transformers>=4.31.0
torch>=2.0.0

# Data querying
pandasql>=0.7.3

# Evaluation metrics
rouge>=1.0.1
nltk>=3.8.0

# Experiment tracking
wandb>=0.15.0

# Visualization
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
